create separate SpriteSymbolInline component
